+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   [ResourcesTreeNode](https://chird.github.io/meteoJS/doc/module-meteoJS_modelviewer_resourcesTreeNode.ResourcesTreeNode.html).
 - Method `getSelectedVariable` in
   [Container](https://chird.github.io/meteoJS/doc/module-meteoJS_modelviewer_container.Container.html).
+### Changed
+- Internal data structure in [Node](https://chird.github.io/meteoJS/doc/module-meteoJS_modelviewer_node.Node.html).
 
 ## [1.12.0] - 2021-03-01
 ### Added